refactor: eliminate brightray::BrowserMainParts (#15336)
