+++ atom/browser/atom_browser_main_parts_mac.mm
@@ -7,6 +7,8 @@
 #include "atom/browser/mac/atom_application_delegate.h"
 #include "base/mac/bundle_locations.h"
 #include "base/mac/foundation_util.h"
+#include "base/path_service.h"
+#include "brightray/browser/brightray_paths.h"
 #include "ui/base/l10n/l10n_util_mac.h"
 
 namespace atom {
@@ -16,7 +18,7 @@ void AtomBrowserMainParts::PreMainMessageLoopStart() {
   AtomApplicationDelegate* delegate = [[AtomApplicationDelegate alloc] init];
   [NSApp setDelegate:delegate];
 
-  brightray::BrowserMainParts::PreMainMessageLoopStart();
+  PreMainMessageLoopStartCommon();
 
   // Prevent Cocoa from turning command-line arguments into
   // |-application:openFiles:|, since we already handle them directly.
@@ -28,6 +30,34 @@ void AtomBrowserMainParts::PreMainMessageLoopStart() {
 void AtomBrowserMainParts::FreeAppDelegate() {
   [[NSApp delegate] release];
   [NSApp setDelegate:nil];
+}
+
+void AtomBrowserMainParts::OverrideAppLogsPath() {
+  base::FilePath path;
+  NSString* bundleName =
+      [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleName"];
+  NSString* logsPath =
+      [NSString stringWithFormat:@"Library/Logs/%@", bundleName];
+  NSString* libraryPath =
+      [NSHomeDirectory() stringByAppendingPathComponent:logsPath];
+
+  base::PathService::Override(brightray::DIR_APP_LOGS,
+                              base::FilePath([libraryPath UTF8String]));
+}
+
+// Replicates NSApplicationMain, but doesn't start a run loop.
+void AtomBrowserMainParts::InitializeMainNib() {
+  auto infoDictionary = base::mac::OuterBundle().infoDictionary;
+
+  auto principalClass =
+      NSClassFromString([infoDictionary objectForKey:@"NSPrincipalClass"]);
+  auto application = [principalClass sharedApplication];
+
+  NSString* mainNibName = [infoDictionary objectForKey:@"NSMainNibFile"];
+  auto mainNib = [[NSNib alloc] initWithNibNamed:mainNibName
+                                          bundle:base::mac::FrameworkBundle()];
+  [mainNib instantiateWithOwner:application topLevelObjects:nil];
+  [mainNib release];
 }
 
 }  // namespace atom
